fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	int64_t a, b, c, temp;
+	long long int a, b, c, temp;
 
 	b = 1;
 	c = 2;
